Fix (hopefully) all Import Cell ITMs/ITPOs
1. The Force Hide Land flags
Flags can have nonsense values outside of the actual flag masks.
Also renamed them to land_flags in the records since that field handles
more in FO4. Once we land the Invent split, we should return to this and
rename C.ForceHideLand to C.LandFlags.

2. Inexact float comparisons
Had to introduce a new game constants for this, which is not going to
scale to other patchers. The right solution is somewhere in the records,
same as with case-insensitive strings.